Adds a lot of working screen:
Posting a tweet, replying to a tweet, navigating inside the app between profiles and tweets by clicking on profile pictures, tweets... Comment count and like count working. There might be work to do on the cache system for it to be more invisible to the user. The app is working quite well and functions well with the server.